Rename `Types` module to `types`.
