+++ trnsysGUI/StorageTank.py
@@ -20,7 +20,7 @@ from trnsysGUI.MyQFileSystemModel import MyQFileSystemModel
 from trnsysGUI.MyQTreeView import MyQTreeView
 from trnsysGUI.PortItem import PortItem
 from trnsysGUI.TeePiece import TeePiece
-from trnsysGUI.Types.createType1924 import Type1924_TesPlugFlow
+from trnsysGUI.types.createType1924 import Type1924_TesPlugFlow
 import trnsysGUI.images as _img
